Add files via upload
The aim of this assignment1 is to familiarize you with writing x86 code. Covered arithmetic, conditional, and looping components in x86-64  also be able to convert a given program ina higher level programming language into its equivalent assembly code.
The aim of this assignment2 is to familiarize you with procedure calls, conditional jumps and stacks.